refactor: 1급객체 blocks를 생성하라
